Refactor TabsPage using react-tabs library
